Prefer not to pass judgement on conferences
Remove quotes around "no budget". Quotation marks can imply sarcasm
about the quoted content, which isn't our intention. We want to open a
dialog without initial bias.

Many conferences are legitimately unable to pay speakers' travel costs
or to offer honorarium. The budgetary concerns faced by organizers,
especially of conferences in their first or second year, are big and
scary. The first year, organizers likely have no idea what they're doing
and the large costs associated with a conference are intimidating.

It is not uncommon for the costs of the event to outweigh the revenue
generated by tickets and sponsors. It's very easy to underestimate the
costs when pricing tickets and to not be able to attract enough sponsor
money to make up the difference.

In a conference's second year, organizers are faced by the chance that
the positive feedback they received for their ostensibly successful
first year was disingenuous. This can lead to not selling enough tickets
and an even harder time attracting sponsors (as well as speakers).

We understand that these concerns are a mitigating factor in offering to
pay the way of speakers.